refactor/가격 데이터타입 전부 int로 변경,db컬럼도 전부 int로 통일

+++ src/main/java/ko/dh/goot/order/dto/OrderResponse.java
@@ -9,6 +9,6 @@ import lombok.Setter;
 @AllArgsConstructor
 public class OrderResponse {
     private Long orderId;
-    private long expectedAmount; // 서버가 최종 확정한 금액
+    private int expectedAmount; // 서버가 최종 확정한 금액
 
 }
